putting ui tests into separate file and ignoring for jenkins
